+++ docs-new/source/index.rst
@@ -1,0 +1,20 @@
+.. VRS documentation master file, created by
+   sphinx-quickstart on Sun Sep 26 12:48:05 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to VRS's documentation!
+===============================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
