--- docs-old/source/index.rst
@@ -1,29 +1,0 @@
-GA4GH Variation Representation Specification
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-The Variation Representation Specification (VRS, pronounced "verse") is a standard
-developed by the Global Alliance for Genomic Health to facilitate and
-improve sharing of genetic information.  The Specification consists of
-a JSON Schema for representing many classes of genetic variation,
-conventions to maximize the utility of the schema, and a Python
-implementation that promotes adoption of the standard.
-
-.. admonition:: Citation
-
-   **The GA4GH Variation Representation Specification (VRS):
-   a Computational Framework for the Precise Representation and
-   Federated Identification of Molecular Variation.**
-   Wagner AH, Babb L, Alterovitz G, Baudis M, Brush M, Cameron DL, ..., Hart RK. 
-   bioRxiv. 2021. `doi:10.1101/2021.01.15.426843 <https://www.biorxiv.org/content/10.1101/2021.01.15.426843v1>`__
-
-
-.. toctree::
-   :maxdepth: 2
-   :includehidden:
-
-   introduction
-   terms_and_model
-   schema
-   impl-guide/index
-   releases/index
-   appendices/index
