+++ docs/Makefile
@@ -1,10 +1,9 @@
 # Minimal makefile for Sphinx documentation
 #
 
-# You can set these variables from the command line, and also
-# from the environment for the first two.
-SPHINXOPTS    ?=
-SPHINXBUILD   ?= sphinx-build
+# You can set these variables from the command line.
+SPHINXOPTS    =
+SPHINXBUILD   = sphinx-build
 SOURCEDIR     = source
 BUILDDIR      = build
 
@@ -21,4 +20,4 @@ help:
 
 
 watch:
-	while true; do find "$(SOURCEDIR)" -name \*.rst | entr -dn make html; done
+	while true; do find source -name \*.rst | entr -dn make html; done
