+++ docs/source/_static/vrs.yaml
