+++ docs/source/appendices/equivalence.rst
