+++ docs/source/appendices/truncated_digest_collision_analysis.rst
