+++ docs/source/images/normalize.png
