+++ docs/source/releases/1.0.rst
