+++ docs/source/releases/1.2.rst
