+++ docs/source/terms_and_model.rst
